Move url launcher to utils
